Rewrite CV generation script to ts
